+++ climaf/standard_operators.py
@@ -17,38 +17,45 @@ def load_standard_operators():
     #
     # Compute scripts
     #
+    cscript('select' ,scriptpath+'mcdo.sh "" ${out} ${var} ${period_iso} ${domain} ${ins} ',
+            commuteWithTimeConcatenation=True, commuteWithSpaceConcatenation=True)
+    #
     cscript('ccdo',
-            scriptpath+'mcdo.sh ${operator} ${out} ${var} ${period_iso} ${domain} ${ins}' ,
-            doSqueezeSpace=True,doSqueezeTime=True)
+            scriptpath+'mcdo.sh ${operator} ${out} ${var} ${period_iso} ${domain} ${ins}')
     #
     cscript('space_average',
             scriptpath+'mcdo.sh fldmean ${out} ${var} ${period_iso} ${domain} ${ins}', 
-            doSqueezeSpace=True)
+            commuteWithTimeConcatenation=True)
     #
     cscript('time_average' ,
             scriptpath+'mcdo.sh timavg  ${out} ${var} ${period_iso} ${domain} ${ins}' ,
-            doSqueezeTime=True)
+            commuteWithSpaceConcatenation=True)
     #
     cscript('llbox' ,
             scriptpath+'mcdo.sh ""  ${out} ${var} ${period_iso} '
-            '${latmin},${latmax},${lonmin},${lonmax} ${ins}')
+            '${latmin},${latmax},${lonmin},${lonmax} ${ins}',
+            commuteWithTimeConcatenation=True, commuteWithSpaceConcatenation=True)
     #
     cscript('regrid' ,
-            scriptpath+'regrid.sh ${in} ${in_2} ${out} ${option}')
+            scriptpath+'regrid.sh ${in} ${in_2} ${out} ${option}',
+            commuteWithTimeConcatenation=True, commuteWithSpaceConcatenation=True)
     #
     cscript('regridn' ,
-            scriptpath+'regrid.sh ${in} ${cdogrid} ${out} ${option}')
+            scriptpath+'regrid.sh ${in} ${cdogrid} ${out} ${option}',
+            commuteWithTimeConcatenation=True, commuteWithSpaceConcatenation=True)
     #
     cscript('mean_and_std',
             scriptpath+'mean_and_std.sh ${in} ${var} ${out} ${out_sdev}', 
             # This tells CliMAF how to name output 'sdev' using input variable name
             sdev_var="std(%s)" , 
-            doSqueezeSpace=True)
+            commuteWithTimeConcatenation=True)
     #
     # Declare plot scripts
-    cscript('ncview'    ,'ncview ${in} 1>/dev/null 2>&1&' , format=None)
+    cscript('ncview'    ,'ncview ${in} 1>/dev/null 2>&1&' )
+    #
     cscript('timeplot'  , scriptpath+'timeplot.sh ${in} ${out} ${var} ${title}',format="png")
+    #
     cscript('plotmap'   , "(ncl -Q "+ scriptpath +"plotmap.ncl infile=${in} "
             "plotname=${out} cmap=${color} vmin=${min} vmax=${max} vdelta=${delta} "
-            "var=${var} title=${crs} scale=${scale} offset=${offset} units=${units} ;"
-            "convert ${out} -trim ${out})",format="png")
+            "var=${var} title=${crs} scale=${scale} offset=${offset} units=${units}) "
+            "; convert ${out} -trim ${out}", format="png")